Avoid creation of irrelevant temporary files on Windows
