+++ util/binfmt_misc/check_windows.go
@@ -3,8 +3,13 @@
 package binfmt_misc
 
 import (
+	"errors"
 	"os/exec"
 )
 
 func withChroot(cmd *exec.Cmd, dir string) {
+}
+
+func check(bin string) error {
+	return errors.New("binfmt is not supported on Windows")
 }
